Consolodate the Byron CLI parsers
Follow the same module structure as we're using for Shelley.

All the commands now live in Cardano.CLI.Byron.Commands and all the CLI
parsers in Cardano.CLI.Byron.Parsers. The "common" stuff was all Byron
specific so that moves there too.

Had to renest the command types so that we have the top level
ClientCommand has just Byron and Shelley (and --verison) and everything
else is now nested within ByronCommand, where previously it was top
level.

Despite renesting the types, we keep the same CLI command hierarchy for
now.
