+++ cardano-cli/src/Cardano/CLI/Byron/Run.hs
@@ -1,5 +1,5 @@
 module Cardano.CLI.Byron.Run
-  ( runByronClientCommand
+  ( runNodeCmd
   ) where
 
 import           Cardano.Prelude
@@ -8,9 +8,6 @@ import           Cardano.CLI.Byron.Parsers
 import           Cardano.CLI.Byron.UpdateProposal
 import           Cardano.CLI.Byron.Vote (runVoteCreation, submitByronVote)
 import           Cardano.CLI.Ops
-
-runByronClientCommand :: ByronCommand -> ExceptT CliError IO ()
-runByronClientCommand (NodeCmd cmd) = runNodeCmd cmd
 
 runNodeCmd :: NodeCmd -> ExceptT CliError IO ()
 runNodeCmd (CreateVote configFp sKey upPropFp voteBool outputFp) =
